Fixed a problem with moveRelativeDegreex10k() not working. We should have been passing a signed integer. Also, not the cast to a signed int inside the inline function definition. Without that, the calculation is incorrect. Tested on hardware.

+++ stepper.cpp
@@ -65,7 +65,7 @@ void stepper::testMoveAbsolute(int positionNew) {
 
 void stepper::testMoveAbsoluteDegree(int positionNewDegree) {
     oss() << endl << "testMoveAbsoluteDegree START" << endl;
-    testMoveAbsolute((int)(positionNewDegree * 10000 / degreesPerMicrostepx10k()));
+    testMoveAbsolute((int)(positionNewDegree * 10000 / (int)degreesPerMicrostepx10k()));
 }
 
 void stepper::testVelocityMove(int v) {
